docs/Networking/examples/graphql_server/requirements.txt
strawberry-graphql[fastapi]>=0.200.0
fastapi>=0.100.0
uvicorn>=0.23.0
